Refactor code to extract Ui, Parser, TaskList class
