+++ src/main/java/Ui.java
@@ -1,0 +1,15 @@
+public class Ui {
+    private final String line = "_____________________________________________________";
+
+    public void greet() {
+        System.out.println(line);
+        System.out.println("Hello! I'm Gerald_Bot\n" + "What can I do for you?");
+        System.out.println(line);
+    }
+
+    public void bye() {
+        System.out.println(line);
+        System.out.println("Bye. Hope to see you again soon!");
+        System.out.println(line);
+    }
+}
